+++ compiler-plugin/src/main/kotlin/tech/mappie/ir/resolving/classes/sources/ExplicitClassMappingSource.kt
@@ -16,7 +16,7 @@ data class ExplicitPropertyMappingSource(
     override val transformation : PropertyMappingTransformation?,
     val forceNonNull: Boolean,
 ) : ExplicitClassMappingSource, TransformableClassMappingSource {
-    val getterType =  reference.getter!!.owner.returnType
+    val getterType = (reference.type as IrSimpleType).arguments.last().typeOrFail
     override val type = type(getterType.let { if (forceNonNull) it.makeNotNull() else it }, transformation)
     override val origin = reference
 }