+++ compiler-plugin/src/main/kotlin/tech/mappie/ir/resolving/classes/sources/ImplicitClassMappingSource.kt
@@ -24,18 +24,20 @@ data class ParameterDefaultValueMappingSource(
 
 data class ImplicitPropertyMappingSource(
     val property: IrProperty,
+    val propertyType: IrType,
     val parameter: Name,
     val parameterType: IrType,
     override val transformation: PropertyMappingTransformation?,
 ) : ImplicitClassMappingSource, TransformableClassMappingSource {
-    override val type = type(property.getter!!.returnType, transformation)
+    override val type = type(propertyType, transformation)
 }
 
 data class FunctionMappingSource(
     val function: IrFunction,
+    val functionType: IrType,
     val parameter: Name,
     val parameterType: IrType,
     override val transformation: PropertyMappingTransformation?,
 ) : ImplicitClassMappingSource, TransformableClassMappingSource {
-    override val type = type(function.returnType, transformation)
+    override val type = type(functionType, transformation)
 }