+++ compiler-plugin/src/main/kotlin/tech/mappie/ir/util/TypeUtils.kt
@@ -1,13 +1,13 @@
 package tech.mappie.ir.util
 
+import org.jetbrains.kotlin.ir.declarations.IrTypeParametersContainer
 import org.jetbrains.kotlin.ir.types.*
-
-import org.jetbrains.kotlin.ir.util.classId
-import org.jetbrains.kotlin.ir.util.fqNameWhenAvailable
-import org.jetbrains.kotlin.ir.util.parentAsClass
-import org.jetbrains.kotlin.ir.util.isSubtypeOf
+import org.jetbrains.kotlin.ir.util.*
 import org.jetbrains.kotlin.ir.util.isNullable
+import org.jetbrains.kotlin.ir.util.isSubtypeOf
+
 import org.jetbrains.kotlin.name.StandardClassIds.Annotations.FlexibleNullability
+import tech.mappie.exceptions.MappiePanicException.Companion.panic
 import tech.mappie.ir.MappieIrRegistrar.Companion.context
 
 fun IrType.isMappableFrom(other: IrType): Boolean = when {
@@ -37,3 +37,11 @@ fun IrType.hasFlexibleNullabilityAnnotation(): Boolean =
 fun IrType.isPrimitive(): Boolean =
     isBoolean() || isByte() || isShort() || isInt() || isLong() || isChar() || isFloat() || isDouble()
             || isStringClassType() || isUByte() || isUShort() || isUInt() || isULong() || isNumber()
+
+fun IrType.substituteTypeVariable(container: IrTypeParametersContainer, arguments: List<IrTypeArgument>) =
+    if (isTypeParameter()) {
+        val mapping = container.typeParameters.zip(arguments).toMap()
+        mapping[classifierOrNull!!.owner]?.typeOrNull ?: panic("Could not resolve generic type", container)
+    } else {
+        this
+    }