+++ website/src/changelog.md
@@ -6,6 +6,7 @@ changelog:
     title: "v2.1.20-1.2.3"
     items:
       - "[#173](https://github.com/Mr-Mappie/mappie/issues/173) fixed compilation error on multiplatform where the this receiver parameter was generated incorrectly."
+      - "[#195](https://github.com/Mr-Mappie/mappie/issues/195) fixed issues with both generic sources and target."
   - date: "2025-04-09"
     title: "v2.1.20-1.2.2"
     items: